added validation on the empty fields on gym class update

+++ 210_project/gymClass.cs
@@ -14,8 +14,8 @@ namespace _210_project
         }
 
         SqlConnection connection = new SqlConnection(Utility.connection);
+
         DataTable dt = new DataTable();
-        int index;
 
         private void dataShow()
         {
@@ -51,11 +51,36 @@ namespace _210_project
         private void gymClass_Load(object sender, EventArgs e)
         {
             dataShow();
-            index = 0;
+            gymClassesDataGridView_Click(null, null);
+            searchByComBox.SelectedText = "Class ID";
+        }
+
+        private bool validate_update()
+        {
+            bool filled = false;
+            string require = "* Required";
+
+            if (titleTxtBox.Text.Trim() != "" && categoryTxtBox.Text.Trim() != "" && nosTxtBox.Text.Trim() != "") filled = true;
+
+            if (titleTxtBox.Text.Trim() == "") titleRequiredLbl.Text = require;
+            if (categoryTxtBox.Text.Trim() == "") categoryRequiredLbl.Text = require;
+            if (nosTxtBox.Text.Trim() == "") nosRequiredLbl.Text = require;
+
+            // have to check the dates
+            // start date cannot be smaller than the end date
+
+            return filled;
         }
 
         private void updatebtn_Click(object sender, EventArgs e)
         {
+            // validation
+            if (!validate_update())
+            {
+                return;
+            }
+            
+
             if (classIDLbl.Text != "")
             {
                 SqlConnection connection = new SqlConnection(Utility.connection);
@@ -69,13 +94,48 @@ namespace _210_project
                 MessageBox.Show("Updated");
                 dataShow();
             }
-            else MessageBox.Show("No Data to update.");
+            else MessageBox.Show("No Data to update.", "No Data!");
             
         }
 
         private void searchBtn_Click(object sender, EventArgs e)
         {
+            if (searchTxtBox.Text.Trim() == "")
+            {
+                searchRequiredLbl.Text = "* Required";
+                return;
+            }
 
+
+            SqlCommand command = connection.CreateCommand();
+            command.CommandType = CommandType.Text;
+            command.CommandText = "SELECT * FROM gym_classes WHERE ";
+
+            switch (searchByComBox.Text.Trim())
+            {
+                case "Class ID":
+                    command.CommandText += "id LIKE '%" + searchTxtBox.Text + "%'";
+                    break;
+
+                case "Title":
+                    command.CommandText += "title LIKE '%" + searchTxtBox.Text + "%'";
+                    break;
+
+                case "Category":
+                    command.CommandText += "category LIKE '%" + searchTxtBox.Text + "%'";
+                    break;
+            }
+
+            connection.Open();
+            command.ExecuteNonQuery();
+            connection.Close();
+            SqlDataAdapter adapter = new SqlDataAdapter(command);
+            dt.Clear();
+            adapter.Fill(dt);
+            gymClassesDataGridView.DataSource = dt;
+
+            // calling the click function to show the data
+            gymClassesDataGridView_Click(null, null);
         }
 
         private void addBtn_Click(object sender, EventArgs e)
@@ -88,6 +148,7 @@ namespace _210_project
         void gym_class_inserted(object sender, EventArgs e)
         {
             dataShow();
+            gymClassesDataGridView_Click(null, null);
         }
 
         private void gymClassesDataGridView_Click(object sender, EventArgs e)
@@ -105,18 +166,27 @@ namespace _210_project
 
         private void deleteBtn_Click(object sender, EventArgs e)
         {
-            var confirm = MessageBox.Show("The data will be deleted permanently.", "Confirm Delete", MessageBoxButtons.YesNo);
-            if (confirm == DialogResult.Yes)
+            if (classIDLbl.Text != "")
             {
-                SqlCommand command = connection.CreateCommand();
-                command.CommandType = CommandType.Text;
-                command.CommandText = "DELETE FROM gym_classes WHERE id = '" + classIDLbl.Text + "'";
-                connection.Open();
-                command.ExecuteNonQuery();
-                connection.Close();
-                dataShow();
-                gymClassesDataGridView_Click(null, null);
+                var confirm = MessageBox.Show("The data will be deleted permanently.", "Confirm Delete", MessageBoxButtons.YesNo);
+                if (confirm == DialogResult.Yes)
+                {
+                    SqlCommand command = connection.CreateCommand();
+                    command.CommandType = CommandType.Text;
+                    command.CommandText = "DELETE FROM gym_classes WHERE id = '" + classIDLbl.Text + "'";
+                    connection.Open();
+                    command.ExecuteNonQuery();
+                    connection.Close();
+                    dataShow();
+                    gymClassesDataGridView_Click(null, null);
+                }
             }
+            else MessageBox.Show("There is not data to delete.", "No Data!");
+        }
+
+        private void searchTxtBox_TextChanged(object sender, EventArgs e)
+        {
+            if (searchTxtBox.Text.Trim() == "") dataShow();
         }
     }
 }
